feat: improve Pydantic V2 compat checks
