+++ src/polus/plugins/_plugins/_compat.py
@@ -1,0 +1,4 @@
+"""Compat of Pydantic."""
+import pydantic
+
+PYDANTIC_V2 = pydantic.VERSION.startswith("2.")
