+++ src/polus/plugins/_plugins/classes/__init__.py
@@ -1,32 +1,16 @@
 """Plugin classes and functions."""
-import pydantic
 
-PYDANTIC_VERSION = pydantic.__version__
-
-if PYDANTIC_VERSION.split(".")[0] == "1":
-    from polus.plugins._plugins.classes.plugin_classes_v1 import PLUGINS
-    from polus.plugins._plugins.classes.plugin_classes_v1 import ComputePlugin
-    from polus.plugins._plugins.classes.plugin_classes_v1 import Plugin
-    from polus.plugins._plugins.classes.plugin_classes_v1 import _load_plugin
-    from polus.plugins._plugins.classes.plugin_classes_v1 import get_plugin
-    from polus.plugins._plugins.classes.plugin_classes_v1 import list_plugins
-    from polus.plugins._plugins.classes.plugin_classes_v1 import load_config
-    from polus.plugins._plugins.classes.plugin_classes_v1 import refresh
-    from polus.plugins._plugins.classes.plugin_classes_v1 import remove_all
-    from polus.plugins._plugins.classes.plugin_classes_v1 import remove_plugin
-    from polus.plugins._plugins.classes.plugin_classes_v1 import submit_plugin
-elif PYDANTIC_VERSION.split(".")[0] == "2":
-    from polus.plugins._plugins.classes.plugin_classes_v2 import PLUGINS
-    from polus.plugins._plugins.classes.plugin_classes_v2 import ComputePlugin
-    from polus.plugins._plugins.classes.plugin_classes_v2 import Plugin
-    from polus.plugins._plugins.classes.plugin_classes_v2 import _load_plugin
-    from polus.plugins._plugins.classes.plugin_classes_v2 import get_plugin
-    from polus.plugins._plugins.classes.plugin_classes_v2 import list_plugins
-    from polus.plugins._plugins.classes.plugin_classes_v2 import load_config
-    from polus.plugins._plugins.classes.plugin_classes_v2 import refresh
-    from polus.plugins._plugins.classes.plugin_classes_v2 import remove_all
-    from polus.plugins._plugins.classes.plugin_classes_v2 import remove_plugin
-    from polus.plugins._plugins.classes.plugin_classes_v2 import submit_plugin
+from polus.plugins._plugins.classes.plugin_classes import PLUGINS
+from polus.plugins._plugins.classes.plugin_classes import ComputePlugin
+from polus.plugins._plugins.classes.plugin_classes import Plugin
+from polus.plugins._plugins.classes.plugin_classes import _load_plugin
+from polus.plugins._plugins.classes.plugin_classes import get_plugin
+from polus.plugins._plugins.classes.plugin_classes import list_plugins
+from polus.plugins._plugins.classes.plugin_classes import load_config
+from polus.plugins._plugins.classes.plugin_classes import refresh
+from polus.plugins._plugins.classes.plugin_classes import remove_all
+from polus.plugins._plugins.classes.plugin_classes import remove_plugin
+from polus.plugins._plugins.classes.plugin_classes import submit_plugin
 
 __all__ = [
     "Plugin",
