+++ src/polus/plugins/_plugins/classes/plugin_classes.py
@@ -10,11 +10,12 @@ from typing import Any
 from typing import Optional
 from typing import Union
 
+from polus.plugins._plugins._compat import PYDANTIC_V2
 from polus.plugins._plugins.classes.plugin_base import BasePlugin
-from polus.plugins._plugins.io.io_v2 import DuplicateVersionFoundError
-from polus.plugins._plugins.io.io_v2 import Version
-from polus.plugins._plugins.io.io_v2 import _in_old_to_new
-from polus.plugins._plugins.io.io_v2 import _ui_old_to_new
+from polus.plugins._plugins.io._io import DuplicateVersionFoundError
+from polus.plugins._plugins.io._io import Version
+from polus.plugins._plugins.io._io import _in_old_to_new
+from polus.plugins._plugins.io._io import _ui_old_to_new
 from polus.plugins._plugins.manifests import InvalidManifestError
 from polus.plugins._plugins.manifests import _load_manifest
 from polus.plugins._plugins.manifests import validate_manifest
@@ -87,14 +88,22 @@ def list_plugins() -> list:
 
 
 def _get_config(plugin: Union["Plugin", "ComputePlugin"], class_: str) -> dict:
-    model_ = json.loads(plugin.model_dump_json())
-    model_["_io_keys"] = deepcopy(plugin._io_keys)  # type: ignore
+    if PYDANTIC_V2:
+        model_ = json.loads(plugin.model_dump_json())
+        model_["_io_keys"] = deepcopy(plugin._io_keys)  # type: ignore
+    else:
+        # ignore mypy if pydantic < 2.0.0
+        model_ = plugin.dict()  # type: ignore
     # iterate over I/O to convert to dict
     for io_name, io in model_["_io_keys"].items():
-        model_["_io_keys"][io_name] = json.loads(io.model_dump_json())
-        # overwrite val if enum
-        if io.type.value == "enum":
-            model_["_io_keys"][io_name]["value"] = io.value.name  # str
+        if PYDANTIC_V2:
+            model_["_io_keys"][io_name] = json.loads(io.model_dump_json())
+            # overwrite val if enum
+            if io.type.value == "enum":
+                model_["_io_keys"][io_name]["value"] = io.value.name  # str
+        elif io["type"] == "enum":  # pydantic V1
+            val_ = io["value"].name  # mapDirectory.raw
+            model_["_io_keys"][io_name]["value"] = val_.split(".")[-1]  # raw
     for inp in model_["inputs"]:
         inp["value"] = None
     model_["class"] = class_
@@ -114,7 +123,15 @@ class Plugin(WIPPPluginManifest, BasePlugin):
     """
 
     id: uuid.UUID  # noqa: A003
-    model_config = ConfigDict(extra="allow", frozen=True)
+    if PYDANTIC_V2:
+        model_config = ConfigDict(extra="allow", frozen=True)
+    else:
+
+        class Config:  # pylint: disable=R0903
+            """Config."""
+
+            extra = "allow"
+            allow_mutation = False
 
     def __init__(self, _uuid: bool = True, **data: dict) -> None:
         """Init a plugin object from manifest."""
@@ -123,7 +140,13 @@ class Plugin(WIPPPluginManifest, BasePlugin):
         else:
             data["id"] = uuid.UUID(str(data["id"]))  # type: ignore
 
+        if not PYDANTIC_V2:  # pydantic V1
+            data["version"] = cast_version(data["version"])
+
         super().__init__(**data)
+
+        if not PYDANTIC_V2:  # pydantic V1
+            self.Config.allow_mutation = True
 
         self._io_keys = {i.name: i for i in self.inputs}
         self._io_keys.update({o.name: o for o in self.outputs})
@@ -202,7 +225,15 @@ class ComputePlugin(ComputeSchema, BasePlugin):
         save_manifest(path): save plugin manifest to specified path
     """
 
-    model_config = ConfigDict(extra="allow", frozen=True)
+    if PYDANTIC_V2:
+        model_config = ConfigDict(extra="allow", frozen=True)
+    else:  # pydantic V1
+
+        class Config:  # pylint: disable=R0903
+            """Config."""
+
+            extra = "allow"
+            allow_mutation = False
 
     def __init__(
         self,
@@ -346,7 +377,11 @@ def submit_plugin(
     org_path.mkdir(exist_ok=True, parents=True)
     if not org_path.joinpath(out_name).exists():
         with org_path.joinpath(out_name).open("w", encoding="utf-8") as file:
-            manifest_ = json.loads(plugin.model_dump_json())
+            if not PYDANTIC_V2:  # pydantic V1
+                manifest_ = plugin.dict()  # type: ignore
+                manifest_["version"] = manifest_["version"]["version"]
+            else:  # PYDANTIC V2
+                manifest_ = json.loads(plugin.model_dump_json())
             json.dump(manifest_, file, indent=4)
 
     # Refresh plugins list
@@ -372,7 +407,9 @@ def get_plugin(
     """
     if version is None:
         return _load_plugin(PLUGINS[name][max(PLUGINS[name])])
-    return _load_plugin(PLUGINS[name][Version(version)])
+    if PYDANTIC_V2:
+        return _load_plugin(PLUGINS[name][Version(version)])
+    return _load_plugin(PLUGINS[name][Version(**{"version": version})])  # Pydantic V1
 
 
 def load_config(config: Union[dict, Path, str]) -> Union[Plugin, ComputePlugin]:
@@ -412,7 +449,13 @@ def remove_plugin(plugin: str, version: Optional[Union[str, list[str]]] = None) 
             for version_ in version:
                 remove_plugin(plugin, version_)
             return
-        version_ = Version(version) if not isinstance(version, Version) else version
+        if not PYDANTIC_V2:  # pydantic V1
+            if not isinstance(version, Version):
+                version_ = cast_version(version)
+            else:
+                version_ = version
+        else:  # pydanitc V2
+            version_ = Version(version) if not isinstance(version, Version) else version
         path = PLUGINS[plugin][version_]
         path.unlink()
     refresh()
